# hyp-instrument/instrument-api/src/main/resources/static/vcp-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "description": "DL Virtual Captive Platform REST API Integration with External Systems",
    "version": "1.0.0",
    "title": "VCP REST API Integration"
  },
...
  "servers": [
    {
      "url": "https://webleasingapi.deutsche-leasing.com/api/v1"
    }
  ],
  "paths": {
    "/oauth/token": {
      "post": {
        "operationId": "getToken",
        "summary": "Get token",
        "description": "Get a bearer token by providing the client id and the client secret from the External System.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/GetTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetTokenRequest": {
        "type": "object",
        "required": [
          "grant_type",
          "client_id",
          "client_secret"
        ],
        "properties": {
          "grant_type": {
            "description": "Supported grant type is *client_credentials*.",
            "nullable": false,
            "type": "string",
            "example": "client_credentials"
          },
          "client_id": {
            "nullable": false,
            "type": "string"
          },
          "client_secret": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "GetTokenResponse": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "nullable": false
          },
          "token_type": {
            "type": "string",
            "nullable": false
          },
          "expires_in": {
            "type": "integer",
            "format": "int32",
            "nullable": false
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "oauth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "scopes": {}
          }
        }
      }
    }
  }
}